pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "accord-core"
version = "0.4.0"
description = "Governance-first multi-agent orchestration toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
  {name = "Accord Collective", email = "ops@accord.ai"},
]
dependencies = [
  "pyyaml>=6",
  "jsonschema>=4.18",
  "cryptography>=41",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]
llm = [
  "openai>=1.3.0",
]

[project.scripts]
accord-orchestrator = "orchestrator.runtime:main"
accord-experiments = "orchestrator.run_experiment:main"
accord-prov = "scripts.provtools:cli"
accord-policy-pipeline = "scripts.policy_synth_pipeline:cli"
accord-validate-alou = "scripts.validate_alou:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestrator", "mcp", "scripts"]
